Allow disabling plugins via the githooks.plugin option
We have one option to enable plugins (githooks.plugin) and another
option to disable them (githooks.disable). Since disable has higher
priority, we can enable a plugin a disable it later but we cannot do the
reverse. In order to make it possible to enable and disable plugins in
any order we have to have a single list of plugins, where each item
tells if a plugin is being enabled or disabled.

So, we're deprecating the githooks.disable option and extending the
githooks.plugin option with the ability to disable a plugin by prefixing
its name with an exclamation point.

This is a backwards compatible change. In a future version we're going
to remove the githooks.disable option.