Use 'undef' to reset a configuration option
Configuration options can take a list of values, which are set from
higher to lower levels of configuration. Sometimes one wants to reset in
a lower level (e.g. local) all values set in higher levels (e.g. system
or global). Now there is a general way to do it. You only need to set
the option to the string 'undef', which is taken as a special mark
meaning to reset the option configuration, as if it hadn't been
configured previously.