Use warn/die instead of carp/croak
When reporting checking errors to the user, the fail_on_faults routine
was using the carp/croak methods, which always shows the filename and
line number where the error occurs, even when the error message ends in
a newline.

We don't need this information because it always shows the same place in
which the routine is invoked.

Thanks @nugged for bringing this to my attention. [0]

Ref0: https://github.com/gnustavo/Git-Hooks/commit/6795420aa3dccca66f55eee69ada3793d4bd001d#r113136608