Fix char decoding in get_config
When the CONFIG_ENCODING variable was set and the git configuration was
empty, the get_config method would try to decode an undefined value.

The fix also allowed us to avoid a conditional, which eliminated one
level of nesting, which required a "no critic" directive to placate
Perl::Critic. That's not necessary anymore.